Edit 20241211 1 (#7)
* Update 2024-12-11

Fixed issues in the UI and wall building mechanism

* Update 20241217

New features
- Overhaul of the preview system to allow for dynamic addition of points between preexisting points, as well as the removal of points

Bug Fixes
- Fixes to the Floor Mapping System

* Update 20241222

- Added Door Feature
- Fixed Wall Intersection Bug where Wall Will Move Microscopically
- Codebase Optimisations

+++ Assets/Scripts/BuildMode/PlacementData.cs
@@ -7,12 +7,12 @@ public class PlacementData
 {
     public GameObject prefab { get; private set; }
     public Vector3 occupiedPosition;
-    public int rotation;
+    public float rotation;
     public Vector2Int size { get; private set; }
     public int ID { get; private set; }
     //public int PlacedObjectIndex { get; private set; }
 
-    public PlacementData(GameObject prefab, Vector3 occupiedPosition, int rotation, Vector2Int size, int iD)
+    public PlacementData(GameObject prefab, Vector3 occupiedPosition, float rotation, Vector2Int size, int iD)
     {
         this.prefab = prefab;
         this.occupiedPosition = occupiedPosition;
